Move current search to advanced search
Resolves: #15776
