+++ app/javascript/components/explore/classes.vue
@@ -60,7 +60,7 @@ import 'd3'
 import 'jquery'
 import { initSetup } from 'lib/explore/classes'
 import { mapActions, mapGetters } from 'vuex'
-import SemesterInput from 'components/search/SearchFormSemesterInput'
+import SemesterInput from 'components/shared/SemesterInput'
 
 export default {
   components: {
