+++ app/javascript/components/explore/universe.vue
@@ -43,7 +43,7 @@ import 'd3'
 import 'jquery'
 import { initSetup } from 'lib/explore/universe'
 import { mapActions, mapGetters } from 'vuex'
-import SemesterInput from 'components/search/SearchFormSemesterInput'
+import SemesterInput from 'components/shared/SemesterInput'
 
 export default {
   components: {
