+++ app/javascript/components/search/advanced/BasicSearchFieldWeightDropdown.vue
