+++ app/javascript/components/search/advanced/SearchResultRow.vue
