+++ app/javascript/views/SearchBasic.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="py-5">
+    <h3>Search:</h3>
+  </div>
+</template>
